libstagefright: Make it possible to skip OMX buffer reallocation (2)
Some devices don't like the call to setParameter() at this point, so
skip this call if enough buffers are already allocated. This check
was present in KitKat but got removed when code to allocate extra-
buffers was introduced.

This is activated only for omap4 for now.

Based on, and requires http://review.cyanogenmod.org/#/c/79137/

This patch skips another possible buffer reallocation, which
can be fatal on these devices.